// apps/web/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      restaurants: {
        Row: {
          address: string
          booking_url: string | null
          created_at: string | null
          cuisine_types: string[] | null
          features: string[] | null
          google_place_id: string | null
          google_rating: number | null
          id: string
          is_verified: boolean | null
          location: unknown | null
          menu_url: string | null
          name: string
          name_chinese: string | null
          opening_hours: Json | null
          openrice_rating: number | null
          phone: string | null
          photos: Json | null
          price_range: number | null
          updated_at: string | null
        }
        Insert: {
          address: string
          booking_url?: string | null
          created_at?: string | null
          cuisine_types?: string[] | null
          features?: string[] | null
          google_place_id?: string | null
          google_rating?: number | null
          id?: string
          is_verified?: boolean | null
          location?: unknown | null
          menu_url?: string | null
          name: string
          name_chinese?: string | null
          opening_hours?: Json | null
          openrice_rating?: number | null
          phone?: string | null
          photos?: Json | null
          price_range?: number | null
          updated_at?: string | null
        }
        Update: {
          address?: string
          booking_url?: string | null
          created_at?: string | null
          cuisine_types?: string[] | null
          features?: string[] | null
          google_place_id?: string | null
          google_rating?: number | null
          id?: string
          is_verified?: boolean | null
          location?: unknown | null
          menu_url?: string | null
          name?: string
          name_chinese?: string | null
          opening_hours?: Json | null
          openrice_rating?: number | null
          phone?: string | null
          photos?: Json | null
          price_range?: number | null
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}